Pipeline refinements: naming, docs, primary constructors, IDE0004 analyzer (#28)
- Rename NodeAllocator → ChainNodeAllocator with its own partial file
- Move NodeMutation to its own partial file
- Rename IProducer.Bootstrap → CreateInitialPayload
- Document IConsumer.Consume and frameStartNanoseconds
- Extract ConsumptionLoop deferred consumer state into nested DeferredConsumerScheduler
- Convert ProductionLoop to primary constructor with explicit field initializers
- Thread real CancellationToken through ProductionLoop.Bootstrap
- Enable IDE0004 analyzer to flag redundant casts, fix all 6 instances

+++ Engine/Program.cs
@@ -14,7 +14,7 @@ Console.CancelKeyPress += (_, eventArgs) => { eventArgs.Cancel = true; cancellat
 // 0..N-1, relying on the OS to time-share when both are active.
 var workerCount = Math.Max(1, Environment.ProcessorCount);
 
-var saveIntervalNanoseconds = (long)SimulationConstants.SaveIntervalTicks
+var saveIntervalNanoseconds = SimulationConstants.SaveIntervalTicks
                               * SimulationConstants.TickDurationNanoseconds;
 
 SimulationEngine.Create(
